\name{extractoR.github-package}
\alias{extractoR.github-package}
\alias{extractoR.github}
\docType{package}
\encoding{UTF-8}
\title{Github R packages metadata extractor}
\description{
  extractoR.github is part of extractoR which is a set of R packages
  that can be used to fetch, extract and query CRAN packages metadata.
  extractoR.github parses CRAN web pages, downloads and extracts
  packages.
}

\author{
Maëlick Claes <himself@maelick.net> & Prateek Rastogi <p.rastogi@outlook.com>

Maintainer: Maëlick Claes <himself@maelick.net>
}
\keyword{fetch}
\keyword{github}
